Revert "extract createAssetNodesFromData to its own file"
This reverts commit 544a3c39f427715b36ea04d84743346e16a0030a.
